+++ agent/temboardagent/cli/app.py
@@ -2,12 +2,13 @@ import logging
 import os
 import datetime
 import getpass
-import sys
 from argparse import _VersionAction
-from platform import python_version
 from socket import getfqdn
 from textwrap import dedent
 
+from ..core import workers
+from ..discover import Discover, inspect_versions
+from ..queries import QUERIES
 from ..toolkit.configuration import OptionSpec
 from ..toolkit.errors import UserError
 from ..web import HTTPDService
@@ -18,11 +19,6 @@ from ..toolkit.configuration import MergedConfiguration
 from ..toolkit.proctitle import ProcTitleManager
 from ..toolkit.signing import load_public_key
 from ..toolkit.tasklist.sqlite3_engine import TaskListSQLite3Engine
-from ..toolkit.versions import (
-    format_pq_version,
-    read_distinfo,
-    read_libpq_version,
-)
 from ..notification import NotificationMgmt
 from ..version import __version__
 
@@ -78,6 +74,7 @@ class TemboardAgentApplication(BaseApplication):
             app=self, setproctitle=setproctitle, name='worker pool',
             task_queue=task_queue, event_queue=event_queue)
         self.services.append(self.worker_pool)
+        self.worker_pool.add(workers)
 
         self.scheduler = taskmanager.SchedulerService(
             app=self, setproctitle=setproctitle, name='scheduler',
@@ -91,6 +88,9 @@ class TemboardAgentApplication(BaseApplication):
         self.bootstrap(args=args, environ=environ, service=command.is_service)
         self.log_versions()
         config = self.config
+
+        self.discover = Discover(self)
+        self.discover.read()
 
         # TaskList engine setup must be done before we load the plugins
         self.scheduler.task_list_engine = TaskListSQLite3Engine(
@@ -108,6 +108,8 @@ class TemboardAgentApplication(BaseApplication):
         if config.postgresql.instance:
             setproctitle.prefix += config.postgresql.instance + ': '
 
+        QUERIES.load()
+
         self.start_datetime = datetime.datetime.now()
         self.reload_datetime = None
         self.pid = os.getpid()
@@ -122,7 +124,8 @@ class TemboardAgentApplication(BaseApplication):
 
     def apply_config(self):
         self.postgres = Postgres(app=self, **self.config.postgresql)
-
+        self.postgres.connection_lost_observers.append(
+            self.discover.connection_lost)
         return super().apply_config()
 
     def bootstrap_plugins(self):
@@ -189,7 +192,7 @@ class TemboardAgentApplication(BaseApplication):
                 yield spec
 
     def log_versions(self):
-        versions = VersionAction.inspect_versions()
+        versions = inspect_versions()
         logger.debug(
             "Running on %s %s.",
             versions['distname'], versions['distversion'])
@@ -242,29 +245,8 @@ class VersionAction(_VersionAction):
     """)
 
     def __call__(self, parser, *_):
-        print((self.fmt % self.inspect_versions()).strip())
+        print((self.fmt % inspect_versions()).strip())
         parser.exit()
-
-    @classmethod
-    def inspect_versions(cls):
-        from bottle import __version__ as bottle_version
-        from psycopg2 import __version__ as psycopg2_version
-        from cryptography import __version__ as cryptography_version
-
-        distinfos = read_distinfo()
-
-        return dict(
-            temboard=__version__,
-            temboardbin=sys.argv[0],
-            psycopg2=psycopg2_version,
-            python=python_version(),
-            pythonbin=sys.executable,
-            bottle=bottle_version,
-            distname=distinfos['NAME'],
-            distversion=distinfos['VERSION'],
-            libpq=format_pq_version(read_libpq_version()),
-            cryptography=cryptography_version,
-        )
 
 
 def list_options_specs():