+++ agent/temboardagent/plugins/dashboard/__init__.py
@@ -74,34 +74,9 @@ def dashboard_hostname():
     return metrics.get_hostname(default_app().temboard.config)
 
 
-@bottle.get('/os_version')
-def dashboard_os_version():
-    return metrics.get_os_version()
-
-
-@bottle.get('/pg_version')
-def dashboard_pg_version(pgconn):
-    return metrics.get_pg_version(pgconn)
-
-
-@bottle.get('/n_cpu')
-def dashboard_n_cpu():
-    return metrics.get_n_cpu()
-
-
 @bottle.get('/databases')
 def dashboard_databases(pgconn):
     return metrics.get_databases(pgconn)
-
-
-@bottle.get('/info')
-def dashboard_info(pgconn):
-    return metrics.get_info(pgconn, default_app().temboard.config)
-
-
-@bottle.get('/max_connections')
-def dashboard_max_connections(pgconn):
-    return metrics.get_max_connections(pgconn)
 
 
 @workers.register(pool_size=1)