+++ agent/temboardagent/queries/__init__.py
@@ -1,0 +1,3 @@
+from ..toolkit.queries import QueryFiler
+
+QUERIES = QueryFiler(__path__[0])